Convert the dropbox_app resource to a custom resource
