+++ libraries/resource_dropbox_app_mac_os_x.rb
@@ -1,0 +1,63 @@
+# encoding: utf-8
+# frozen_string_literal: true
+#
+# Cookbook Name:: dropbox
+# Library:: resource_dropbox_app_mac_os_x
+#
+# Copyright 2014-2017, Jonathan Hartman
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+require_relative 'resource_dropbox_app'
+
+class Chef
+  class Resource
+    # A Chef resource for Dropbox packages for Mac OS X.
+    #
+    # @author Jonathan Hartman <j@p4nt5.com>
+    class DropboxAppMacOsX < Resource::DropboxApp
+      PATH ||= '/Applications/Dropbox.app'.freeze
+
+      provides :dropbox_app, platform_family: 'mac_os_x'
+
+      #
+      # Use a dmg_package resource to install the Dropbox app.
+      #
+      action :install do
+        s = source_path
+        dmg_package 'Dropbox' do
+          source s
+          volumes_dir 'Dropbox Installer'
+        end
+      end
+
+      #
+      # In the absence of an uninstall script for OS X, kill any running
+      # instance of Dropbox and delete its directories.
+      #
+      action :remove do
+        execute 'killall Dropbox' do
+          ignore_failure true
+        end
+        [::File.expand_path('/Library/DropboxHelperTools'),
+         ::File.expand_path(PATH)].each do |d|
+          directory d do
+            recursive true
+            action :delete
+          end
+        end
+      end
+    end
+  end
+end
